refactor(tegg): rename utils.ts to AgentStoreUtils.ts for naming consistency
Follow tegg convention where utility files use PascalCase
(e.g., ObjectUtils.ts, TimerUtil.ts, NameUtil.ts).
