+++ tegg/core/agent-runtime/test/utils.test.ts
@@ -2,7 +2,7 @@ import { strict as assert } from 'node:assert';
 
 import { describe, it } from 'vitest';
 
-import { newMsgId, nowUnix } from '../src/utils.ts';
+import { newMsgId, nowUnix } from '../src/AgentStoreUtils.ts';
 
 describe('core/agent-runtime/test/utils.test.ts', () => {
   it('nowUnix should return current unix timestamp in seconds', () => {
